requirements.txt
# Core Flask dependencies
Flask==2.0.1
Werkzeug==2.0.1
click==7.1.2
itsdangerous==2.0.1
Jinja2==3.0.1

# Database
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
pyodbc==4.0.39
alembic==1.7.1

# Authentication
Flask-Login==0.5.0
bcrypt==3.2.0

# Forms
Flask-WTF==0.15.1
WTForms==2.3.3
email-validator==1.1.3

# Utils
python-dotenv==0.19.0
pytz==2021.1

# Security
Flask-Talisman==0.8.1

# Testing and Development
pytest==6.2.5
coverage==5.5